[#9] Rename chrome_extension_helper.js to chrome.js
