+++ test/helpers/chrome.js
