+++ test/utils/display_manager.js
@@ -1,6 +1,6 @@
 import DisplayManager from '../../src/utils/display_manager';
 import assert from 'power-assert';
-import createMockChromeObject from '../helpers/chrome_extension_helper';
+import createMockChromeObject from '../helpers/chrome';
 import createDisplay from '../fixtures/display';
 import createWindow from '../fixtures/window';
